Refactor plugin module structure
