+++ src-plugin/src/plugin/state_support.rs
@@ -1,0 +1,85 @@
+use std::sync::Arc;
+
+use serde::{Deserialize, Serialize};
+use wrac_clap_adapter::{PluginError, PluginResult, PluginState, PluginStateSupport};
+
+use crate::gui::GuiStateNotifier;
+use crate::plugin::{PARAM_BYPASS_ID, PARAM_GAIN_ID};
+use crate::state::{
+    EditorPage, ParameterStateSnapshot, ProjectState, ProjectStateStore, SharedState,
+};
+
+/// DAW project に保存する plugin state の serialize 形式 (JSON)。
+///
+/// realtime parameter は [`SharedState`] から、editor-only state は
+/// [`ProjectStateStore`] から snapshot し、この 1 形式に合成して host へ渡す。
+#[derive(Debug, Serialize, Deserialize)]
+pub(crate) struct SavedPluginState {
+    pub(crate) gain: f32,
+    #[serde(default)]
+    pub(crate) bypass: bool,
+    #[serde(default)]
+    pub(crate) editor_page: EditorPage,
+}
+
+pub(super) struct WracGainStateSupport {
+    project_state: Arc<ProjectStateStore>,
+    shared: Arc<SharedState>,
+    gui_notifier: Arc<GuiStateNotifier>,
+}
+
+impl WracGainStateSupport {
+    pub(super) fn new(
+        project_state: Arc<ProjectStateStore>,
+        shared: Arc<SharedState>,
+        gui_notifier: Arc<GuiStateNotifier>,
+    ) -> Self {
+        Self {
+            project_state,
+            shared,
+            gui_notifier,
+        }
+    }
+}
+
+// project 保存で `save_state`、復元で `restore_state`。bytes 形式は自由なので、
+// デバッグしやすい JSON にしている。
+impl PluginStateSupport for WracGainStateSupport {
+    fn save_state(&self) -> PluginResult<PluginState> {
+        let project = self.project_state.snapshot();
+        let params = self.shared.snapshot_parameters();
+        log::debug!(
+            "saving plugin state: gain={}, bypass={}, editor_page={}",
+            params.gain,
+            params.bypass,
+            project.editor_page.as_str()
+        );
+        let bytes = serde_json::to_vec(&SavedPluginState {
+            gain: params.gain,
+            bypass: params.bypass,
+            editor_page: project.editor_page,
+        })
+        .map_err(|_| PluginError::InvalidState)?;
+        Ok(PluginState { bytes })
+    }
+
+    fn restore_state(&self, state: PluginState) -> PluginResult<()> {
+        log::debug!("restoring plugin state: byte_count={}", state.bytes.len());
+        let state: SavedPluginState =
+            serde_json::from_slice(&state.bytes).map_err(|_| PluginError::InvalidState)?;
+        let project = ProjectState {
+            editor_page: state.editor_page,
+        };
+        self.project_state.commit(project);
+        self.shared.restore_parameters(ParameterStateSnapshot {
+            gain: state.gain,
+            bypass: state.bypass,
+        });
+        self.gui_notifier
+            .notify_parameter(PARAM_GAIN_ID, self.shared.gain());
+        self.gui_notifier
+            .notify_parameter(PARAM_BYPASS_ID, f32::from(self.shared.bypass()));
+        self.gui_notifier.notify_editor_page(project.editor_page);
+        Ok(())
+    }
+}
